test: move e2e helper functions to utils_test.go
